lib: Rename dir and file http_gw/http_utils.py (HTTP->REST)
