+++ pytest_tests/tests/services/rest_gate/test_rest_object.py
@@ -11,7 +11,7 @@ from helpers.rest_gate import (
 )
 from helpers.neofs_verbs import put_object_to_random_node
 from helpers.wellknown_acl import PUBLIC_ACL
-from http_gw.http_utils import get_object_and_verify_hashes
+from rest_gw.rest_utils import get_object_and_verify_hashes
 from neofs_env.neofs_env_test_base import NeofsEnvTestBase
 
 logger = logging.getLogger("NeoLogger")
