+++ pytest_tests/tests/services/rest_gate/test_rest_system_header.py
@@ -16,7 +16,7 @@ from helpers.rest_gate import (
 )
 from helpers.neofs_verbs import get_netmap_netinfo, get_object_from_random_node, head_object
 from helpers.wellknown_acl import PUBLIC_ACL
-from http_gw.http_utils import get_object_and_verify_hashes
+from rest_gw.rest_utils import get_object_and_verify_hashes
 from neofs_env.neofs_env_test_base import NeofsEnvTestBase
 
 logger = logging.getLogger("NeoLogger")
